Experiment with projectile and barel explosions

+++ Assets/GameFiles/Scripts/Singletons/GameManager.cs
@@ -257,9 +257,9 @@ public class GameManager : MonoBehaviour
     }
 
     //TODO: ponder
-    public int GetScoreFromDamage(int damage)
+    public int GetScoreFromDamage(float damage)
     {
-        return damage / 2;
+        return (int)damage / 2;
     }
 
     //TODO: ponder
